Separate map hasher from rfc6962 hasher
Logs do not need to compute empty branches.
Rather than creating a unified hasher for logs and maps, create separate
hashers that will panic when used in the wrong setting.
