+++ merkle/map_verifier_test.go
@@ -17,12 +17,12 @@ package merkle
 import (
 	"testing"
 
-	"github.com/google/trillian/merkle/rfc6962"
+	"github.com/google/trillian/merkle/maphasher"
 	"github.com/google/trillian/testonly"
 )
 
 func TestVerifyMap(t *testing.T) {
-	h := rfc6962.DefaultHasher
+	h := maphasher.Default
 	tv := mapInclusionTestVector[0]
 
 	// Copy the bad proof so we don't mess up the good proof.
