+++ merkle/maphasher/generic.go
@@ -12,57 +12,61 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package rfc6962 provides hashing functionality according to RFC6962.
-package rfc6962
+// Package maphasher provides hashing for maps.
+package maphasher
 
 import (
 	"crypto"
-	_ "crypto/sha256" // SHA256 is the default algorithm.
 	"fmt"
 )
 
 // Domain separation prefixes
 const (
-	RFC6962LeafHashPrefix = 0
-	RFC6962NodeHashPrefix = 1
+	leafHashPrefix = 0
+	nodeHashPrefix = 1
 )
 
-// DefaultHasher is a SHA256 based TreeHasher.
-var DefaultHasher = New(crypto.SHA256)
+// Default is a SHA256 based TreeHasher for maps.
+var Default = New(crypto.SHA256)
 
 // TreeHasher implements the RFC6962 tree hashing algorithm.
 // Empty branches within the tree are plain interior nodes e1 = H(e0, e0) etc.
 type TreeHasher struct {
 	crypto.Hash
+	nullHashes [][]byte
 }
 
 // New creates a new TreeHasher on the passed in hash function.
 func New(h crypto.Hash) *TreeHasher {
 	m := &TreeHasher{Hash: h}
+	m.nullHashes = m.createNullHashes()
 	return m
 }
 
 // String returns a string representation for debugging.
 func (t *TreeHasher) String() string {
-	return fmt.Sprintf("rfc6962Hash{%v}", t.Hash)
+	return fmt.Sprintf("genericHash{%v}", t.Hash)
 }
 
-// EmptyRoot returns a special case for an empty tree.
+// EmptyRoot returns the root of an empty tree.
 func (t *TreeHasher) EmptyRoot() []byte {
-	return t.New().Sum(nil)
+	return t.HashEmpty(t.Size() * 8)
 }
 
 // HashEmpty returns the hash of an empty branch at a given depth.
 // A depth of 0 indictes the hash of an empty leaf.
 func (t *TreeHasher) HashEmpty(depth int) []byte {
-	panic("HashEmpty() is not implemented for rfc6962 hasher")
+	if depth < 0 || depth >= len(t.nullHashes) {
+		panic(fmt.Sprintf("HashEmpty(%v) out of bounds", depth))
+	}
+	return t.nullHashes[depth]
 }
 
 // HashLeaf returns the Merkle tree leaf hash of the data passed in through leaf.
 // The data in leaf is prefixed by the LeafHashPrefix.
 func (t *TreeHasher) HashLeaf(leaf []byte) []byte {
 	h := t.New()
-	h.Write([]byte{RFC6962LeafHashPrefix})
+	h.Write([]byte{leafHashPrefix})
 	h.Write(leaf)
 	return h.Sum(nil)
 }
@@ -71,7 +75,7 @@ func (t *TreeHasher) HashLeaf(leaf []byte) []byte {
 // The hashed structure is NodeHashPrefix||l||r.
 func (t *TreeHasher) HashChildren(l, r []byte) []byte {
 	h := t.New()
-	h.Write([]byte{RFC6962NodeHashPrefix})
+	h.Write([]byte{nodeHashPrefix})
 	h.Write(l)
 	h.Write(r)
 	return h.Sum(nil)
