+++ merkle/rfc6962/rfc6962_test.go
@@ -15,64 +15,9 @@ package rfc6962
 
 import (
 	"bytes"
-	"crypto"
-	"encoding/base64"
 	"encoding/hex"
 	"testing"
-
-	"github.com/google/trillian/merkle"
 )
-
-const (
-	// Expected root hash of an empty sparse Merkle tree.
-	// This was taken from the C++ SparseMerkleTree tests in
-	// github.com/google/certificate-transparency.
-	emptyMapRootB64 = "xmifEIEqCYCXbZUz2Dh1KCFmFZVn7DUVVxbBQTr1PWo="
-)
-
-func TestEmptyRoot(t *testing.T) {
-	emptyRoot, err := base64.StdEncoding.DecodeString(emptyMapRootB64)
-	if err != nil {
-		t.Fatalf("couldn't decode empty root base64 constant.")
-	}
-	mh := New(crypto.SHA256)
-	rootLevel := mh.Size() * 8
-	if got, want := mh.HashEmpty(rootLevel), emptyRoot; !bytes.Equal(got, want) {
-		t.Fatalf("HashEmpty(0): %x, want %x", got, want)
-	}
-}
-
-// Compares the old HStar2 empty branch algorithm to the new.
-func TestHStar2Equivalence(t *testing.T) {
-	m := New(crypto.SHA256)
-	star := hstar{
-		hasher:          m,
-		hStarEmptyCache: [][]byte{m.HashLeaf([]byte(""))},
-	}
-	fullDepth := m.Size() * 8
-	for i := 0; i < fullDepth; i++ {
-		if got, want := m.HashEmpty(i), star.hStarEmpty(i); !bytes.Equal(got, want) {
-			t.Errorf("HashEmpty(%v): \n%x, want: \n%x", i, got, want)
-		}
-	}
-}
-
-// Old hstar2 empty cache algorithm.
-type hstar struct {
-	hasher          merkle.MapHasher
-	hStarEmptyCache [][]byte
-}
-
-// hStarEmpty calculates (and caches) the "null-hash" for the requested tree level.
-// Note: here level 0 is the leaf and level 255 is the root.
-func (s *hstar) hStarEmpty(n int) []byte {
-	if len(s.hStarEmptyCache) <= n {
-		emptyRoot := s.hStarEmpty(n - 1)
-		h := s.hasher.HashChildren(emptyRoot, emptyRoot)
-		s.hStarEmptyCache = append(s.hStarEmptyCache, h)
-	}
-	return s.hStarEmptyCache[n]
-}
 
 func TestRfc6962Hasher(t *testing.T) {
 	hasher := DefaultHasher
